refactor: CustomException name 필드 복구
